Normalize card creditLimit from cents to reais in GET /cards response

+++ src/main/java/com/example/smart_wallet/modules/card/infrastructure/web/dto/GetCardDetailsResponseDTO.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record GetCardDetailsResponseDTO(
         UUID id,
         String name,
-        Integer creditLimit,
+        BigDecimal creditLimit,
         BigDecimal currentInvoice,
         String dueDateLabel,
         String status
